cmake: Do not force installation of travex targets
This seems like a helpful behavior when travex is included as a submodule but
the parent project does not want to build travex libraries by 'make all'. In
such cases the build followed by `make install` will not choke due to missing
travex libraries. This does not change the default behavior.